+++ multiboot2/Changelog.md
@@ -6,6 +6,7 @@
 - **BREAKING** Fixed wrong creation of `EFIMemoryMapTag`.
   `EFIMemoryMapTag::new` was replaced by `EFIMemoryMapTag::new_from_descs` and
   `EFIMemoryMapTag::new_from_map`.
+- `ModuleTag::new`'s `end` parameter now must be bigger than `start`.
 
 ## 0.20.0 (2024-05-01)